Efectos de estado y union de Aventureros con el sistema de pelea

+++ Assets/01_Scripts/Adventurers/AdventurerData.cs
@@ -13,8 +13,8 @@ public class AdventurerData
     public GameObject prefab;
     public int level = 1;
 
-    public int maxHealth;
-    public int currentHealth;
+    public float maxHealth;
+    public float currentHealth;
     public int strength;
     public int speed;
     public bool isAlive = true;
